--- apps/hub/src/app/_providers/_config.ts
@@ -1,17 +1,0 @@
-import { createConfig, http } from 'wagmi'
-import { mainnet } from 'wagmi/chains'
-
-export const config = createConfig({
-  chains: [mainnet],
-  ssr: false,
-  transports: {
-    // todo: replace public clients
-    [mainnet.id]: http(),
-  },
-})
-
-declare module 'wagmi' {
-  interface Register {
-    config: typeof config
-  }
-}
