+++ apps/hub/src/app/_providers/index.tsx
@@ -1,4 +1,18 @@
-export { AccountProvider } from './account-provider'
-export { ConnectKitProvider } from './connectkit-provider'
-export { QueryClientProvider } from './query-client-provider'
-export { WagmiProvider } from './wagmi-provider'
+'use client'
+
+import { QueryClientProvider, WagmiProvider } from '@status-im/wallet/providers'
+import { ConnectKitProvider as ConnectKit } from 'connectkit'
+
+interface ProvidersProps {
+  children: React.ReactNode
+}
+
+export function Providers({ children }: ProvidersProps) {
+  return (
+    <WagmiProvider>
+      <QueryClientProvider>
+        <ConnectKit>{children}</ConnectKit>
+      </QueryClientProvider>
+    </WagmiProvider>
+  )
+}
