+++ apps/wallet/src/routes/__root.tsx
@@ -3,6 +3,11 @@
 // import OnboardingPage from '../../../portfolio/src/app/page'
 import { ToastContainer } from '@status-im/components'
 import { Navbar } from '@status-im/wallet/components'
+// import { TanStackRouterDevtools } from '@tanstack/router-devtools'
+// import { NotAllowed } from '../../../portfolio/src/app/_components/not-allowed'
+// import { AccountsProvider } from '../../../portfolio/src/app/_providers/accounts-context'
+// import { ConnectKitProvider } from '../../../portfolio/src/app/_providers/connectkit-provider'
+import { QueryClientProvider, WagmiProvider } from '@status-im/wallet/providers'
 // import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import {
   createRootRouteWithContext,
@@ -12,13 +17,7 @@ import {
   redirect,
 } from '@tanstack/react-router'
 
-// import { TanStackRouterDevtools } from '@tanstack/router-devtools'
-// import { NotAllowed } from '../../../portfolio/src/app/_components/not-allowed'
-// import { AccountsProvider } from '../../../portfolio/src/app/_providers/accounts-context'
-// import { ConnectKitProvider } from '../../../portfolio/src/app/_providers/connectkit-provider'
-import { QueryClientProvider } from '../../../portfolio/src/app/_providers/query-client-provider'
 // import { StatusProvider } from '../../../portfolio/src/app/_providers/status-provider'
-import { WagmiProvider } from '../../../portfolio/src/app/_providers/wagmi-provider'
 import { Link } from '../components/link'
 import { apiClient } from '../providers/api-client'
 import { PendingTransactionsProvider } from '../providers/pending-transactions-context'
