+++ package.json
@@ -33,7 +33,11 @@
   },
   "resolutions": {
     "@types/react": "19.1.0",
-    "@types/react-dom": "19.1.1",
+    "@types/react-dom": "19.1.0",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "viem": "2.29.1",
+    "wagmi": "2.15.2",
     "zod": "3.23.8",
     "vite": "6.3.5",
     "vitest": "3.1.4",
@@ -51,9 +55,9 @@
     "lint-staged": "^15.2.10",
     "prettier": "^3.3.3",
     "prettier-plugin-tailwindcss": "^0.6.6",
-    "react": "^18.3.1",
+    "react": "^19.1.0",
     "react-devtools": "^6.1.5",
-    "react-dom": "^18.3.1",
+    "react-dom": "^19.1.0",
     "rimraf": "^4.4.1",
     "turbo": "^2.3.0",
     "typescript": "^5.6.2",
