+++ packages/wallet/src/config/wagmi.ts
@@ -1,8 +1,10 @@
 import { createConfig, http } from 'wagmi'
 import { mainnet } from 'wagmi/chains'
+import { injected } from 'wagmi/connectors'
 
 export const config = createConfig({
   chains: [mainnet],
+  connectors: [injected()],
   ssr: false,
   transports: {
     // todo: replace public clients
