+++ packages/wallet/src/providers/index.tsx
@@ -1,0 +1,2 @@
+export { QueryClientProvider } from './query-client-provider'
+export { wagmiConfig, WagmiProvider } from './wagmi-provider'
