+++ packages/wallet/src/providers/query-client-provider.tsx
@@ -2,6 +2,8 @@
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
+import type React from 'react'
+
 const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
